Train on real data: fix eval-metric mismatch, scale_pos_weight overfitting, and 3 other bugs found live
First real execution against the downloaded Kaggle dataset (284,807 rows)
surfaced real bugs no amount of synthetic-fixture testing could have
caught -- fixed all of them before trusting the resulting numbers:

- models/champion.py: LightGBM's built-in eval_metric="average_precision"
  string does not compute the same quantity as sklearn's
  average_precision_score under scale_pos_weight -- confirmed by swapping
  in an explicit sklearn-wrapping callable, after which LightGBM's
  internally-reported validation score exactly matched a post-hoc
  recomputation (previously off by ~9x). Early stopping had been silently
  optimizing the wrong objective, producing a champion that scored 0.65 on
  VAL but only 0.13 PR-AUC on TEST -- worse than the baseline.
- configs/train_config.yaml: with the metric fixed, a VAL-only
  hyperparameter search (never touching TEST) found the textbook
  auto-computed scale_pos_weight (~447x) was itself overfitting badly to
  the 350 training-period frauds. scale_pos_weight=1 (no reweighting) with
  moderate regularization generalized far better across every config
  tried. Retrained: TEST PR-AUC 0.130 -> 0.792, now legitimately above the
  0.747 baseline. Also widened the chronological split 60/20/20 ->
  55/20/25 (chosen from fraud counts alone, before training) since the
  original left TEST with only 75 of the required 80 frauds.
- scripts/eda.py: never read train_config.yaml, silently reproducing the
  already-rejected default split; the report's split percentages were also
  hardcoded literals that had gone stale. Both fixed to read from config.
- scripts/build_curated_examples.py: crashed with 'float has no attribute
  sin' -- a row sliced from a mixed-dtype (floats + string transaction_id)
  DataFrame becomes object-dtype, and np.sin() chokes on boxed Python
  floats in an object array. Fixed with an explicit .astype(float).
- scripts/benchmark_latency.py: took minutes for 100 requests against an
  API answering single curl calls in ~10ms -- bare requests.post() opens a
  fresh TCP connection every call; wrapped the loop in a requests.Session()
  for keep-alive, which is also the more realistic thing to measure.

Verified live end-to-end with the corrected model: /score, /explain, and
/chargeback/draft-response all tested against real curated transactions,
including the chargeback safety guard correctly refusing on a real
high-risk transaction and drafting a genuine, evidence-only letter for a
real low-risk one. README/MODEL_CARD/CHALLENGES updated with the real
measured numbers and the full investigation.

+++ scripts/build_curated_examples.py
@@ -56,7 +56,12 @@ def main():
     examples = []
     for category, rows in categories.items():
         for _, row in rows.iterrows():
-            raw_row = row[FEATURE_COLUMNS].to_frame().T
+            # `row` is sliced from `merged`, a mixed-dtype frame (floats +
+            # the string transaction_id column), so a row-wise Series over
+            # it is coerced to object dtype -- .astype(float) is required
+            # here or downstream numpy ufuncs (np.sin in FeatureEngineer)
+            # choke on boxed Python floats inside an object array.
+            raw_row = row[FEATURE_COLUMNS].astype(float).to_frame().T
             contributions = explainer.explain_one(raw_row)
             # Full raw feature vector, shaped exactly like the API's
             # TransactionIn schema, so the dashboard can replay this exact
